requirements.txt
aiohttp
beautifulsoup4
lxml
snscrape
apscheduler
pytz
python-telegram-bot==13.15